Update: Move ModalMobile & ModalTeam under components dir
